+++ rl_envs_forge/envs/labyrinth/display.py
@@ -1,6 +1,6 @@
 import pygame
 
-from rl_envs.envs.labyrinth.constants import *
+from rl_envs_forge.envs.labyrinth.constants import *
 
 
 class Display: