+++ tests/envs/labyrinth/rooms/test_rectangular_room.py
@@ -1,7 +1,7 @@
 import pytest
 import numpy as np
-from rl_envs.envs.labyrinth.room import RectangularRoom
-from rl_envs.envs.labyrinth.constants import *
+from rl_envs_forge.envs.labyrinth.room import RectangularRoom
+from rl_envs_forge.envs.labyrinth.constants import *
 
 
 class TestRectangularRoom: